Removes old tx cursors and entity items used in them
